refactor: 도메인 하위 패키지 생성
