+++ src/main/kotlin/blackjack/README.md
@@ -16,8 +16,8 @@
 - [x] 게임 시작 시 플레이어의 이름을 입력할 수 있다.
 - [x] 게임 시작 시 플레이어의 카드를 출력한다.
 - [x] 게임 시작 후, 각 플레이어는 카드를 뽑을 지 선택할 수 있다.
-- [ ] 플레이어는 카드를 뽑기를 멈출 수 있다.
-- [ ] 플레이어는 숫자 합이 21이 넘으면 카드를 뽑을 수 없다.
+- [x] 플레이어는 카드를 뽑기를 멈출 수 있다.
+- [x] 플레이어는 숫자 합이 21이 넘으면 카드를 뽑을 수 없다.
 - [ ] Ace는 숫자 합에 따라 1 또는 11로 계산할 수 있다.
 - [ ] 카드의 합이 21이하면서, 가까운 플레이어가 승리한다.
