+++ src/main/kotlin/blackjack/domain/Dealer.kt
@@ -1,5 +1,8 @@
 package blackjack.domain
 
+import blackjack.domain.card.Card
+import blackjack.domain.card.CardDeck
+
 class Dealer {
     val cardDeck: CardDeck = CardDeck()
