+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -1,5 +1,7 @@
 package blackjack.domain
 
+import blackjack.domain.card.Card
+
 class Player(
     val name: String = "player"
 ) {
