+++ src/main/kotlin/blackjack/domain/card/CardSuit.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class CardSuit(val forOutput: String) {
     SPADE("스페이드"),
