+++ src/test/kotlin/blackjack/domain/card/CardRankTest.kt
@@ -1,6 +1,6 @@
-package blackjack
+package blackjack.domain.card
 
-import blackjack.domain.CardRank
+import blackjack.domain.card.CardRank
 import io.kotest.matchers.collections.shouldContainAll
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.DisplayName
